Refactor `FactionSystem` integration and update related systems
- Replaced `factionID` with `FactionNames` in multiple components and systems for better typing and readability.
- Updated `PlayerInfo` and `TestingDummyInfo` scriptable objects to use `FactionNames`.
- Refactored `CheckTargetThreatSystem` and `Combinators` to leverage `FactionNames` with relationship logic for affinity calculations.
- Adjusted data handling in `TargetCtx` and `InfluenceComponent` for faction-specific relationships.
- Updated `UniversalRenderPipelineGlobalSettings` and other assets to include additional resources for debugging and rendering.
- Bumped `ProjectVersion` to 6000.2.2f1.

+++ Assets/Systems/IAUS/Components/Components/States/Blob State System/StateBlob.cs
@@ -78,7 +78,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = p.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  }));
              }).Run();
@@ -89,7 +89,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = p.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  }));
              }).Run();
@@ -99,7 +99,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = p.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  }));
              }).Run();
@@ -109,7 +109,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = w.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  }));
              }).Run();
@@ -120,7 +120,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = g.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  })); ;
              }).Run();
@@ -130,7 +130,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = G.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  }));             }).Run();
 
@@ -140,7 +140,7 @@ namespace IAUS.ECS.StateBlobSystem
              //    {
              //        Difficulty = Difficulty.Normal,
              //        aIStates = G.name,
-             //        FactionID = brain.factionID,
+             //        FactionID = (int)brain.FactionID,
              //        NPCLevel = brain.NPCLevel
              //    }); 
             // }).Run();
@@ -149,7 +149,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = G.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  })); ;
              }).Run();
@@ -159,7 +159,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = G.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  }));
              }).Run();
@@ -170,7 +170,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = G.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  })); 
              }).Run();
@@ -181,7 +181,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = G.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  }));
              }).Run();
@@ -192,7 +192,7 @@ namespace IAUS.ECS.StateBlobSystem
                  {
                      Difficulty = Difficulty.Normal,
                      AIStates = G.Name,
-                     FactionID = brain.FactionID,
+                     FactionID = (int)brain.FactionID,
                      NPCLevel = brain.NPCLevel
                  }));
              }).Run();
